creating new dir for data

+++ converter.py
@@ -50,19 +50,20 @@ def raiseError(message, temp_dir, out_dir):
     #sys.exit()
 
 def main(url, out):
-    files = [f for f in os.listdir('./static') if f != 'cookies.txt']
-    for f in files:
-        if (os.path.isfile('./static/' + f)):
-            os.remove('./static/' + f)
-        else:
-            shutil.rmtree('./static/' + f) 
-
+    if False:
+        files = [f for f in os.listdir('./static') if f != 'cookies.txt']
+        for f in files:
+            if (os.path.isfile('./static/' + f)):
+                os.remove('./static/' + f)
+            else:
+                shutil.rmtree('./static/' + f) 
 
     out_dir = out
 #    temp_dir = tempfile.TemporaryDirectory().name
-    temp_dir = "./static"
+    temp_dir = "./temp/" + out_dir
     #out_dir = temp_dir
     print(temp_dir)
+    os.makedirs(temp_dir, exist_ok=True)
 #    get_cookies(temp_dir)
 
     print('getting vid info', url, temp_dir, out_dir)
